+++ Changelog.md
@@ -61,3 +61,7 @@ Stor dag idag.
 
 - Lagt till ett nytt test. Samt ändrat några data-cy och börjar kolla på att få till rätt rating på framsidan.
 - Rättat till mitt test för AllComics komponent så att den fyller ett större syfte och mockar data korrekt.
+
+2024-12-18
+
+- Lagt till code-coverage. Tog en stund att få allt rätt.